+++ src/components/common/GlobalShell.jsx
@@ -1,16 +1,7 @@
 import { useRecoilState } from 'recoil';
-import {
-  AppShell,
-  Header,
-  Container,
-  Button,
-  Text,
-  Burger,
-  useMantineTheme,
-  useMantineColorScheme,
-} from '@mantine/core';
+import { AppShell, Header, Container, Button, Burger, useMantineTheme, useMantineColorScheme } from '@mantine/core';
 import styled from '@emotion/styled';
-import isSideNavOpenState from '../recoil/atom/sideNavOpenedState';
+import isSideNavOpenState from '../../recoil/atom/sideNavOpenedState';
 import SearchBar from './SearchBar';
 import SideNavBar from './SideNavBar';
 
@@ -26,7 +17,7 @@ const Logo = styled.button`
   background-color: transparent;
 `;
 
-const GlobalShell = () => {
+const GlobalShell = ({ children }) => {
   const theme = useMantineTheme();
   const [isOpened, setIsOpend] = useRecoilState(isSideNavOpenState);
   const { colorScheme } = useMantineColorScheme();
@@ -64,7 +55,7 @@ const GlobalShell = () => {
           </SubContainer>
         </Header>
       }>
-      <Text>Resize app to see responsive navbar in action</Text>
+      {children}
     </AppShell>
   );
 };
