+++ src/components/common/Header.jsx
@@ -2,7 +2,7 @@ import { useRecoilState } from 'recoil';
 import styled from '@emotion/styled';
 
 import { Burger, Button, useMantineColorScheme } from '@mantine/core';
-import sideNavOpenedState from '../recoil/atom/sideNavOpenedState';
+import sideNavOpenedState from '../../recoil/atom/sideNavOpenedState';
 
 import SearchBar from './SearchBar';
