+++ src/components/common/index.js
@@ -1,0 +1,5 @@
+export { default as GlobalShell } from './GlobalShell';
+export { default as Header } from './Header';
+export { default as SearchBar } from './SearchBar';
+export { default as SideNavBar } from './SideNavBar';
+export { default as ThemeButton } from './ThemeButton';
