+++ src/components/index.js
@@ -1,10 +1,5 @@
-export { default as GlobalShell } from './GlobalShell';
-export { default as ThemeButton } from './ThemeButton';
 export { default as ListButton } from './ListButton';
 export { default as MyList } from './MyList';
 export { default as Badge } from './Badge';
 export { default as Badges } from './Badges';
 export { default as ListBadges } from './ListBadges';
-export { default as Header } from './Header';
-export { default as SideNavBar } from './SideNavBar';
-export { default as SearchBar } from './SearchBar';
