+++ src/pages/Root.jsx
@@ -1,11 +1,17 @@
 import React from 'react';
-import { GlobalShell, ThemeButton } from '../components';
+import { GlobalShell, ThemeButton } from '../components/common';
+import { Cards, Board } from '../components/rootPage';
 
 const Root = () => {
   console.log('Root');
+
   return (
     <>
-      <GlobalShell>{/* children */}</GlobalShell>
+      <GlobalShell>
+        {/* <Banner /> */}
+        <Board />
+        <Cards />
+      </GlobalShell>
       <ThemeButton />
     </>
   );
